requirements.txt
absl-py==2.2.2
accelerate==1.6.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.3.10
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.3.0
av==15.0.0
babel==2.17.0
basicsr==1.4.2
beautifulsoup4==4.13.4
bleach==6.2.0
blessed==1.21.0
braceexpand==0.1.7
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
clip==0.2.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
deepspeed==0.16.5
defusedxml==0.7.1
diffusers==0.32.2
dill==0.4.0
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.8.1
embreex==2.17.7.post6
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.3.0
ftfy==6.3.1
future==1.0.0
gevent==24.11.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-genai==1.11.0
gpustat==1.1.1
gradio==5.17.0
gradio_client==1.7.1
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
hf-xet==1.1.7
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.34.4
identify==2.6.9
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
iniconfig==2.1.0
inquirerpy==0.3.4
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.34.0
ipywidgets==8.1.7
isoduration==20.11.0
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
kornia==0.8.0
kornia_rs==0.1.8
lazy_loader==0.4
lightning==2.5.1
lightning-utilities==0.14.2
line_profiler==4.2.0
llvmlite==0.44.0
lmdb==1.6.2
lpips==0.1.4
lxml==5.3.1
manifold3d==3.0.1
mapbox_earcut==1.0.3
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.2.0
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nodeenv==1.9.1
notebook==7.4.3
notebook_shim==0.2.4
numba==0.61.2
numpy==1.24.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
open_clip_torch==2.31.0
openai==1.75.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.15.2
pexpect==4.9.0
pfzy==0.3.4
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
plyfile==1.1
pooch==1.8.2
pre_commit==4.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.25.6
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycollada==0.9
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pylance==0.30.0
PyMatting==1.1.13
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-multipart==0.0.20
pytorch-fid==0.3.0
pytorch-lightning==2.5.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==27.0.0
qwen-vl-utils==0.0.11
referencing==0.36.2
regex==2024.11.6
rembg==2.0.65
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
roma==1.5.2.1
rpds-py==0.23.1
rsa==4.9.1
rtree==1.4.0
ruff==0.11.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==2.24.1
setproctitle==1.3.5
shapely==2.0.7
shellingham==1.5.4
shtab==1.7.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
splines==0.3.2
stack-data==0.6.3
starlette==0.46.1
svg.path==6.3
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.20.0a20250503
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.3.13
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.0+cu124
torchaudio==2.5.0+cu124
torchdata==0.9.0
torchmetrics==1.7.0
torchvision==0.20.0+cu124
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.50.0
trimesh==4.6.5
triton==3.1.0
typeguard==4.4.2
typer==0.15.2
types-python-dateutil==2.9.0.20250516
typing_extensions==4.12.2
tyro==0.9.17
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
-e git+https://github.com/facebookresearch/vggt.git@b02cc03ceee70821ed1231a530c1992507ef9862#egg=vggt
vhacdx==0.0.8.post2
virtualenv==20.29.3
viser==0.2.23
wandb==0.19.8
wcwidth==0.2.13
webcolors==24.11.1
webdataset==0.2.111
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xatlas==0.0.9
xformers @ git+https://github.com/facebookresearch/xformers.git@68b7fd14df5eb1d2558c52842b4206a14d2d20e9
xlrd==2.0.1
XlsxWriter==3.2.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
yourdfpy==0.0.57
yt-dlp==2025.4.30
zipp==3.21.0
zope.event==5.0
zope.interface==7.2

